@@ -1,13 +1,11 @@
 {
   "name": "electron-boot",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
+  "private": true,
   "scripts": {
-    "clear:lib": "rimraf packages/*/{lib,dist}",
-    "clear:dependencies": "rimraf packages/*/node_modules && rimraf node_modules",
-    "clear": "rimraf packages/*/{lib,dist,node_modules} && rimraf node_modules && rimraf examples/quick-start-app/node_modules",
-    "test": "jest",
+    "preinstall": "npx only-allow pnpm",
+    "lint": "eslint --cache .",
+    "build": "pnpm -r --filter='./packages/*' run build",
+    "dev": "pnpm -r --parallel --filter='./packages/*' run dev",
     "build:packages": "pnpm --filter './packages/**' build",
     "changeset:beta": "changeset pre enter beta",
     "changeset:init": "changeset init",
@@ -17,25 +15,39 @@
     "release": "pnpm build:packages && pnpm release:only",
     "release:only": "changeset publish --registry=https://registry.npmjs.com/ --access public"
   },
-  "keywords": [],
-  "author": "",
+  "packageManager": "pnpm@9.12.2",
+  "engines": {
+    "node": "^18.0.0 || >=20.0.0"
+  },
   "license": "ISC",
   "devDependencies": {
-    "@changesets/cli": "^2.26.2",
-    "@types/jest": "^29.5.10",
-    "@types/node": "^20.9.4",
+    "@changesets/cli": "^2.27.9",
+    "@eslint/js": "^9.12.0",
+    "@kayahr/jest-electron-runner": "^29.15.0",
+    "@rollup/plugin-json": "^6.1.0",
+    "@rollup/plugin-node-resolve": "^15.3.0",
+    "@rollup/plugin-typescript": "^12.1.1",
+    "@types/jest": "^29.5.13",
+    "@types/node": "^22.7.6",
     "@types/prettier": "^3.0.0",
-    "@types/rimraf": "^4.0.5",
-    "eslint": "^8.54.0",
-    "eslint-config-prettier": "^9.0.0",
+    "eslint": "^9.12.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-n": "^17.11.1",
     "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-prettier": "^5.0.1",
+    "eslint-plugin-prettier": "^5.2.1",
+    "eslint-plugin-regexp": "^2.6.0",
+    "globals": "^15.11.0",
     "jest": "^29.7.0",
-    "prettier": "^3.1.0",
-    "rimraf": "^5.0.5",
-    "ts-jest": "^29.1.1",
-    "ts-node": "^10.9.1",
-    "typedoc": "^0.25.3",
-    "typescript": "^5.3.2"
+    "prettier": "^3.3.3",
+    "rollup": "^4.24.0",
+    "rollup-plugin-dts": "^6.1.1",
+    "rollup-plugin-esbuild": "^6.1.1",
+    "rollup-plugin-node-externals": "^7.1.3",
+    "rollup-plugin-rm": "^1.0.2",
+    "ts-jest": "^29.2.5",
+    "ts-node": "^10.9.2",
+    "typedoc": "^0.26.10",
+    "typescript": "^5.6.3",
+    "typescript-eslint": "^8.9.0"
   }
 }
