@@ -1,18 +1,28 @@
 {
   "name": "@electron-boot/framework",
-  "version": "2.0.9",
+  "version": "2.1.0",
   "description": "fast development framework for electron",
-  "main": "lib/index.js",
-  "typings": "lib/index.d.ts",
+  "main": "dist/cjs/index.cjs",
+  "module": "dist/esm/index.mjs",
+  "types": "dist/types/index.d.ts",
+  "type": "module",
+  "exports": {
+    "types": "./dist/types/index.d.ts",
+    "import": "./dist/esm/index.mjs",
+    "require": "./dist/cjs/index.cjs"
+  },
   "publishConfig": {
     "access": "public"
   },
   "files": [
-    "lib/**/*.js",
-    "lib/**/*.d.ts"
+    "dist"
   ],
   "scripts": {
-    "build": "tsc"
+    "electron": "electron test/fixtures/start.ts",
+    "lint": "eslint --cache --ext .ts -c eslint.config.mjs src/**",
+    "format": "prettier --write --cache --parser typescript \"src/**/*.ts\"",
+    "dev": "rollup -c rollup.config.ts --configPlugin typescript -w",
+    "build": "rollup -c rollup.config.ts --configPlugin typescript "
   },
   "repository": {
     "type": "git",
@@ -31,23 +41,17 @@
   "peerDependencies": {
     "electron": ">=25.8.0",
     "tslib": ">=2.6.2",
-    "typescript": ">=5.2.2"
+    "typescript": ">=5.2.2",
+    "@electron-boot/logger": "workspace:*"
   },
   "dependencies": {
     "@electron-boot/logger": "workspace:*",
-    "fs-extra": "^11.1.1",
-    "lodash": "^4.17.21",
-    "picomatch": "^2.3.1",
-    "reflect-metadata": "^0.1.13"
-  },
-  "engines": {
-    "node": ">=12"
+    "picomatch": "^4.0.2",
+    "radash": "^12.1.0",
+    "reflect-metadata": "^0.2.2"
   },
   "devDependencies": {
-    "@types/fs-extra": "^11.0.1",
-    "@types/lodash": "^4.14.202",
-    "@types/picomatch": "^2.3.0",
-    "electron": "^25.8.0",
-    "rimraf": "^5.0.5"
+    "@types/picomatch": "^3.0.1",
+    "electron": "catalog:electron30"
   }
 }
