@@ -1,18 +1,26 @@
 {
   "name": "@electron-boot/logger",
-  "version": "2.0.6",
+  "version": "2.1.0",
   "description": "fast development framework for electron",
-  "main": "lib/index.js",
-  "typings": "lib/index.d.ts",
+  "main": "dist/cjs/index.cjs",
+  "module": "dist/esm/index.mjs",
+  "types": "dist/types/index.d.ts",
+  "exports": {
+    "types": "./dist/types/index.d.ts",
+    "import": "./dist/esm/index.mjs",
+    "require": "./dist/cjs/index.cjs"
+  },
   "publishConfig": {
     "access": "public"
   },
   "files": [
-    "lib/**/*.js",
-    "lib/**/*.d.ts"
+    "dist"
   ],
   "scripts": {
-    "build": "tsc"
+    "lint": "eslint --cache --ext .ts -c eslint.config.mjs src/**",
+    "format": "prettier --write --cache --parser typescript \"src/**/*.ts\"",
+    "dev": "rollup -c rollup.config.ts --configPlugin typescript -w",
+    "build": "rollup -c rollup.config.ts --configPlugin typescript "
   },
   "repository": {
     "type": "git",
@@ -29,19 +37,14 @@
   "homepage": "https://github.com/electron-boot/electron-boot#readme",
   "license": "MIT",
   "peerDependencies": {
-    "electron": ">=25.8.0",
     "tslib": ">=2.6.2",
     "typescript": ">=5.2.2"
   },
   "dependencies": {
-    "logform": "^2.6.0",
+    "logform": "^2.6.1",
+    "radash": "^12.1.0",
     "winston": "^3.9.0",
-    "winston-daily-rotate-file": "^4.7.1"
-  },
-  "engines": {
-    "node": ">=12"
-  },
-  "devDependencies": {
-    "rimraf": "^5.0.5"
+    "winston-daily-rotate-file": "^5.0.0",
+    "electron": "catalog:electron30"
   }
 }
