@@ -1,60 +1,62 @@
-import { BootstrapOptions, IApplicationContext } from '../interface';
-import { ILogger, LoggerFactory } from '@electron-boot/logger';
-import { inspect } from 'util';
-import { GenericApplicationContext } from '../context/generic.application.context';
+import { inspect } from "node:util";
+import type { ILogger } from "@electron-boot/logger";
+import { LoggerFactory } from "@electron-boot/logger";
+import type { BootstrapOptions, IApplicationContext } from "../interface";
+import { GenericApplicationContext } from "../context/generic.application.context";
 import {
   bindContainer,
   clearBindContainer,
   listPreloadModule,
-} from '../decorators/decorator.manager';
-import { AspectService } from '../service/aspect.service';
-import { EnvironmentService } from '../service/environment.service';
-import { DecoratorService } from '../service/decorator.service';
-import { ConfigService } from '../service/config.service';
-import { SocketService } from '../service/socket.service';
-import { LifecycleService } from '../service/lifecycle.service';
-import { IpcService } from '../service/ipc.service';
-import { EventService } from '../service/event.service';
-import defaultConfig from '../config/config.default';
+} from "../decorators/decorator.manager";
+import { AspectService } from "../service/aspect.service";
+import { EnvironmentService } from "../service/environment.service";
+import { DecoratorService } from "../service/decorator.service";
+import { ConfigService } from "../service/config.service";
+import { SocketService } from "../service/socket.service";
+import { LifecycleService } from "../service/lifecycle.service";
+import { IpcService } from "../service/ipc.service";
+import { EventService } from "../service/event.service";
+import defaultConfig from "../config/config.default";
+
 let stepIdx = 1;
 export class Application {
   protected globalOptions: Partial<BootstrapOptions> = {};
   protected logger: ILogger = LoggerFactory.getLogger(Application);
-  private applicationContext: IApplicationContext;
+  private applicationContext: IApplicationContext | undefined;
 
   private printStepDebugInfo(stepInfo: string) {
     this.logger.debug(`\n\nStep ${stepIdx++}: ${stepInfo}\n`);
   }
 
-  public configure(options: BootstrapOptions = {}) {
+  public configure(options: BootstrapOptions = {}): this {
     this.globalOptions = options;
     return this;
   }
 
   private async prepareGlobalApplicationContext(): Promise<IApplicationContext> {
-    this.printStepDebugInfo('Ready to create applicationContext');
+    this.printStepDebugInfo("Ready to create applicationContext");
 
     this.logger.debug(
-      '[framework]: start "initializeGlobalApplicationContext"'
+      '[framework]: start "initializeGlobalApplicationContext"',
     );
     this.logger.debug(
-      `[framework]: bootstrap options = ${inspect(this.globalOptions)}`
+      `[framework]: bootstrap options = ${inspect(this.globalOptions)}`,
     );
 
     // new applicationContext
-    const applicationContext =
+    const applicationContext: IApplicationContext =
       this.globalOptions.applicationContext ?? new GenericApplicationContext();
 
     // bind moduleStore to DecoratorUtil
-    this.logger.debug('[core]: delegate module map from DecoratorUtil');
+    this.logger.debug("[core]: delegate module map from DecoratorUtil");
     bindContainer(applicationContext);
     this.applicationContext = applicationContext;
 
-    global['ELECTRON_APPLICATION_CONTEXT'] = applicationContext;
+    global["ELECTRON_APPLICATION_CONTEXT"] = applicationContext;
 
-    this.printStepDebugInfo('Ready module detector');
+    this.printStepDebugInfo("Ready module detector");
 
-    this.printStepDebugInfo('Binding inner service');
+    this.printStepDebugInfo("Binding inner service");
     // bind inner service
     applicationContext.bindClass(EnvironmentService);
     applicationContext.bindClass(AspectService);
@@ -65,7 +67,7 @@ export class Application {
     applicationContext.bindClass(LifecycleService);
     applicationContext.bindClass(EventService);
 
-    this.printStepDebugInfo('Binding preload module');
+    this.printStepDebugInfo("Binding preload module");
 
     // bind preload module
     if (
@@ -78,7 +80,7 @@ export class Application {
     }
 
     this.printStepDebugInfo(
-      'Init ConfigService, AspectService, DecoratorService'
+      "Init ConfigService, AspectService, DecoratorService",
     );
     // init default config
     const configService = applicationContext.get(ConfigService);
@@ -95,12 +97,12 @@ export class Application {
     applicationContext.get(DecoratorService, [applicationContext]);
 
     this.printStepDebugInfo(
-      'Load imports(component) and user code configuration module'
+      "Load imports(component) and user code configuration module",
     );
 
     // load import module
     applicationContext.load([].concat(this.globalOptions.imports));
-    this.printStepDebugInfo('Run applicationContext ready method');
+    this.printStepDebugInfo("Run applicationContext ready method");
 
     // bind user code module
     await applicationContext.ready();
@@ -113,14 +115,14 @@ export class Application {
       }
     }
 
-    this.printStepDebugInfo('Load config file');
+    this.printStepDebugInfo("Load config file");
 
     // merge config
     configService.load();
 
     this.logger.debug(
-      '[core]: Current config = %j',
-      configService.getConfiguration()
+      "[core]: Current config = %j",
+      configService.getConfiguration(),
     );
 
     // init ipc
@@ -134,16 +136,16 @@ export class Application {
   private async initializeGlobalApplicationContext(): Promise<IApplicationContext> {
     const applicationContext = await this.prepareGlobalApplicationContext();
 
-    this.printStepDebugInfo('Init runtime');
+    this.printStepDebugInfo("Init runtime");
 
     await applicationContext.getAsync(SocketService, [applicationContext]);
 
-    this.printStepDebugInfo('Init lifecycle');
+    this.printStepDebugInfo("Init lifecycle");
 
     // lifecycle support
     await applicationContext.getAsync(LifecycleService, [applicationContext]);
 
-    this.printStepDebugInfo('Init preload modules');
+    this.printStepDebugInfo("Init preload modules");
 
     // some preload module init
     const modules = listPreloadModule();
@@ -152,12 +154,12 @@ export class Application {
       await applicationContext.getAsync(module);
     }
 
-    this.printStepDebugInfo('End of initialize and start');
+    this.printStepDebugInfo("End of initialize and start");
 
     return applicationContext;
   }
   private async destroyGlobalApplicationContext(
-    applicationContext: IApplicationContext
+    applicationContext: IApplicationContext,
   ) {
     // stop lifecycle
     const lifecycleService =
@@ -167,20 +169,20 @@ export class Application {
     await applicationContext.stop();
     clearBindContainer();
     // clear applicationContext
-    global['ELECTRON_APPLICATION_CONTEXT'] = null;
+    global["ELECTRON_APPLICATION_CONTEXT"] = null;
   }
 
-  public async run() {
+  public async run(): Promise<IApplicationContext> {
     this.applicationContext = await this.initializeGlobalApplicationContext();
     return this.applicationContext;
   }
 
-  public async stop() {
+  public async stop(): Promise<void> {
     if (this.applicationContext) {
       await this.destroyGlobalApplicationContext(this.applicationContext);
     }
   }
   public getApplicationContext(): IApplicationContext {
-    return this.applicationContext;
+    return this.applicationContext!;
   }
 }
