@@ -1,9 +1,10 @@
-import { Application } from './application';
-import { ILogger, LoggerFactory } from '@electron-boot/logger';
-import { BootstrapOptions, IApplicationContext } from '../interface';
+import type { ILogger } from "@electron-boot/logger";
+import { LoggerFactory } from "@electron-boot/logger";
+import type { BootstrapOptions, IApplicationContext } from "../interface";
+import { Application } from "./application";
 
 export class Bootstrap {
-  protected static application: Application;
+  protected static application: Application | null;
   protected static configured = false;
   protected static logger: ILogger = LoggerFactory.getLogger(Bootstrap);
   public static configure(options: BootstrapOptions = {}): typeof Bootstrap {
@@ -19,7 +20,7 @@ export class Bootstrap {
     return this.application;
   }
 
-  public static reset() {
+  public static reset(): void {
     this.configured = false;
     this.application = null;
   }
@@ -27,58 +28,56 @@ export class Bootstrap {
   /**
    * run application
    */
-  public static async run() {
+  public static async run(): Promise<IApplicationContext> {
     if (!this.configured) {
       this.configure();
     }
 
-    process.once('SIGINT', this.onSignal.bind(this, 'SIGINT'));
+    process.once("SIGINT", this.onSignal.bind(this, "SIGINT"));
     // kill(3) Ctrl-\
-    process.once('SIGQUIT', this.onSignal.bind(this, 'SIGQUIT'));
+    process.once("SIGQUIT", this.onSignal.bind(this, "SIGQUIT"));
     // kill(15) default
-    process.once('SIGTERM', this.onSignal.bind(this, 'SIGTERM'));
+    process.once("SIGTERM", this.onSignal.bind(this, "SIGTERM"));
 
-    process.once('exit', this.onExit.bind(this));
+    process.once("exit", this.onExit.bind(this));
 
     this.uncaughtExceptionHandler = this.uncaughtExceptionHandler.bind(this);
-    process.on('uncaughtException', this.uncaughtExceptionHandler);
+    process.on("uncaughtException", this.uncaughtExceptionHandler);
 
     this.unhandledRejectionHandler = this.unhandledRejectionHandler.bind(this);
-    process.on('unhandledRejection', this.unhandledRejectionHandler);
+    process.on("unhandledRejection", this.unhandledRejectionHandler);
 
-    return this.getApplication()
-      .run()
-      .then(value => {
-        this.logger.info('Application started');
-        return value;
-      })
-      .catch(err => {
-        this.logger.error(err);
-        process.exit(1);
-      });
+    try {
+      const applicationContext = await this.getApplication().run();
+      this.logger.info("Application started");
+      return applicationContext;
+    } catch (e) {
+      this.logger.error(e);
+      process.exit(1);
+    }
   }
 
   /**
    * stop application
    */
-  public static async stop() {
+  public static async stop(): Promise<void> {
     await this.getApplication().stop();
-    process.off('uncaughtException', this.uncaughtExceptionHandler);
-    process.off('unhandledRejection', this.unhandledRejectionHandler);
+    process.off("uncaughtException", this.uncaughtExceptionHandler);
+    process.off("unhandledRejection", this.unhandledRejectionHandler);
     this.reset();
   }
   /**
    * on bootstrap receive an exit signal
    * @param signal
    */
-  private static async onSignal(signal) {
-    this.logger.info('receive signal %s, closing', signal);
+  private static async onSignal(signal: NodeJS.Signals) {
+    this.logger.info("receive signal %s, closing", signal);
     try {
       await this.stop();
-      this.logger.info('close done, exiting with code:0');
+      this.logger.info("close done, exiting with code:0");
       process.exit(0);
     } catch (err) {
-      this.logger.error('close with error: ', err);
+      this.logger.error("close with error: ", err);
       process.exit(1);
     } finally {
       LoggerFactory.destroy();
@@ -88,21 +87,21 @@ export class Bootstrap {
    * on bootstrap process exit
    * @param code
    */
-  private static onExit(code) {
-    this.logger.info('exit with code:%s', code);
+  private static onExit(code: NodeJS.Signals) {
+    this.logger.info("exit with code:%s", code);
   }
 
-  private static uncaughtExceptionHandler(err) {
+  private static uncaughtExceptionHandler(err: Error) {
     if (!(err instanceof Error)) {
       err = new Error(String(err));
     }
-    if (err.name === 'Error') {
-      err.name = 'unhandledExceptionError';
+    if (err.name === "Error") {
+      err.name = "unhandledExceptionError";
     }
     this.logger.error(err);
   }
 
-  private static unhandledRejectionHandler(err) {
+  private static unhandledRejectionHandler(err: any) {
     if (!(err instanceof Error)) {
       const newError = new Error(String(err));
       // err maybe an object, try to copy the name, message and stack to the new error instance
@@ -113,8 +112,8 @@ export class Bootstrap {
       }
       err = newError;
     }
-    if (err.name === 'Error') {
-      err.name = 'unhandledRejectionError';
+    if (err.name === "Error") {
+      err.name = "unhandledRejectionError";
     }
     this.logger.error(err);
   }
