@@ -1,2 +1,2 @@
-export * from './application';
-export * from './bootstrap';
+export * from "./application";
+export * from "./bootstrap";
