@@ -1,43 +1,43 @@
 export const KEYS = {
-  OBJECTS_ELEMENT: 'objects',
-  OBJECT_ELEMENT: 'object',
-  IMPORT_ELEMENT: 'import',
-  PROPERTY_ELEMENT: 'property',
-  LIST_ELEMENT: 'list',
-  MAP_ELEMENT: 'map',
-  ENTRY_ELEMENT: 'entry',
-  VALUE_ELEMENT: 'value',
-  PROPS_ELEMENT: 'props',
-  PROP_ELEMENT: 'prop',
-  SET_ELEMENT: 'set',
-  CONSTRUCTOR_ARG_ELEMENT: 'constructor-arg',
-  REF_ELEMENT: 'ref',
-  JSON_ELEMENT: 'json',
-  CONFIGURATION_ELEMENT: 'configuration',
+  OBJECTS_ELEMENT: "objects",
+  OBJECT_ELEMENT: "object",
+  IMPORT_ELEMENT: "import",
+  PROPERTY_ELEMENT: "property",
+  LIST_ELEMENT: "list",
+  MAP_ELEMENT: "map",
+  ENTRY_ELEMENT: "entry",
+  VALUE_ELEMENT: "value",
+  PROPS_ELEMENT: "props",
+  PROP_ELEMENT: "prop",
+  SET_ELEMENT: "set",
+  CONSTRUCTOR_ARG_ELEMENT: "constructor-arg",
+  REF_ELEMENT: "ref",
+  JSON_ELEMENT: "json",
+  CONFIGURATION_ELEMENT: "configuration",
 
-  ID_ATTRIBUTE: 'id',
-  PATH_ATTRIBUTE: 'path',
-  DIRECT_ATTRIBUTE: 'direct',
-  AUTOWIRE_ATTRIBUTE: 'autowire',
-  ASYNC_ATTRIBUTE: 'async',
-  NAME_ATTRIBUTE: 'name',
-  REF_ATTRIBUTE: 'ref',
-  KEY_ATTRIBUTE: 'key',
-  VALUE_ATTRIBUTE: 'value',
-  TYPE_ATTRIBUTE: 'type',
-  EXTERNAL_ATTRIBUTE: 'external',
-  OBJECT_ATTRIBUTE: 'object',
-  RESOURCE_ATTRIBUTE: 'resource',
-  SCOPE_ATTRIBUTE: 'scope',
+  ID_ATTRIBUTE: "id",
+  PATH_ATTRIBUTE: "path",
+  DIRECT_ATTRIBUTE: "direct",
+  AUTOWIRE_ATTRIBUTE: "autowire",
+  ASYNC_ATTRIBUTE: "async",
+  NAME_ATTRIBUTE: "name",
+  REF_ATTRIBUTE: "ref",
+  KEY_ATTRIBUTE: "key",
+  VALUE_ATTRIBUTE: "value",
+  TYPE_ATTRIBUTE: "type",
+  EXTERNAL_ATTRIBUTE: "external",
+  OBJECT_ATTRIBUTE: "object",
+  RESOURCE_ATTRIBUTE: "resource",
+  SCOPE_ATTRIBUTE: "scope",
 
-  ASPECT_ELEMENT: 'aspect',
-  AROUND_ELEMENT: 'around',
-  EXPRESSION_ATTRIBUTE: 'expression',
-  EXECUTE_ATTRIBUTE: 'execute',
+  ASPECT_ELEMENT: "aspect",
+  AROUND_ELEMENT: "around",
+  EXPRESSION_ATTRIBUTE: "expression",
+  EXECUTE_ATTRIBUTE: "execute",
 };
 
-export const FUNCTION_INJECT_KEY = 'midway:function_inject_key';
-export const REQUEST_CTX_KEY = 'ctx';
-export const REQUEST_OBJ_CTX_KEY = '_req_ctx';
-export const CONTAINER_OBJ_SCOPE = '_obj_scope';
+export const FUNCTION_INJECT_KEY = "midway:function_inject_key";
+export const REQUEST_CTX_KEY = "ctx";
+export const REQUEST_OBJ_CTX_KEY = "_req_ctx";
+export const CONTAINER_OBJ_SCOPE = "_obj_scope";
 export const SINGLETON_CONTAINER_CTX = { _MAIN_CTX_: true };
