@@ -1,38 +1,38 @@
-import { GenericApplicationContext } from './generic.application.context';
-import { IApplicationContext } from '../interface';
-import { REQUEST_CTX_KEY } from '../constant';
-import { PIPELINE_IDENTIFIER } from '../decorators/decorator.manager';
+import { GenericApplicationContext } from "./generic.application.context";
+import type { IApplicationContext } from "../interface";
+import { REQUEST_CTX_KEY } from "../constant";
+import { PIPELINE_IDENTIFIER } from "../decorators/decorator.manager";
 
 export class RequestApplicationContext extends GenericApplicationContext {
   private readonly applicationContext: IApplicationContext;
 
-  constructor(ctx, applicationContext: IApplicationContext) {
+  constructor(ctx: any, applicationContext: IApplicationContext) {
     super(applicationContext);
     this.applicationContext = applicationContext;
 
     // update legacy relationship
     this.registry.setIdentifierRelation(
-      this.applicationContext.registry.getIdentifierRelation()
+      this.applicationContext.registry.getIdentifierRelation(),
     );
 
     this.ctx = ctx;
     // register ctx
     this.registerObject(REQUEST_CTX_KEY, ctx);
     // register res
-    this.registerObject('res', {});
+    this.registerObject("res", {});
 
     if (ctx.logger) {
       // register contextLogger
-      this.registerObject('logger', ctx.logger);
+      this.registerObject("logger", ctx.logger);
     }
   }
 
-  init() {
+  init(): void {
     // do nothing
   }
 
-  get<T = any>(identifier: any, args?: any): T {
-    if (typeof identifier !== 'string') {
+  override get<T = any>(identifier: any, args?: any): T {
+    if (typeof identifier !== "string") {
       identifier = this.getIdentifier(identifier);
     }
 
@@ -58,10 +58,11 @@ export class RequestApplicationContext extends GenericApplicationContext {
     if (this.parent) {
       return this.parent.get(identifier, args);
     }
+    return undefined as unknown as T;
   }
 
   async getAsync<T = any>(identifier: any, args?: any): Promise<T> {
-    if (typeof identifier !== 'string') {
+    if (typeof identifier !== "string") {
       identifier = this.getIdentifier(identifier);
     }
 
@@ -87,13 +88,14 @@ export class RequestApplicationContext extends GenericApplicationContext {
     if (this.parent) {
       return this.parent.getAsync<T>(identifier, args);
     }
+    return undefined as unknown as T;
   }
 
-  async ready() {
+  async ready(): Promise<void> {
     // ignore other things
   }
 
-  getContext() {
+  getContext(): any {
     return this.ctx;
   }
 }
