@@ -1,18 +1,18 @@
-import { ObjectIdentifier, PropertyParamDecorator } from '../interface';
-import { isNullOrUndefined, isNumber } from '../utils/types.util';
+import type { ObjectIdentifier, PropertyParamDecorator } from "../interface";
+import { isNullOrUndefined, isNumber } from "../utils/types.util";
 import {
   attachClassMetadata,
   INJECT_CUSTOM_PARAM,
   savePropertyInject,
-} from './decorator.manager';
+} from "./decorator.manager";
 
 export function Autowired(
-  identifier?: ObjectIdentifier
+  identifier?: ObjectIdentifier,
 ): PropertyParamDecorator {
   return (
     target: any,
     propertyKey: string | symbol | undefined,
-    parameterIndex?: number
+    parameterIndex?: number,
   ) => {
     if (isNullOrUndefined(parameterIndex) && propertyKey) {
       savePropertyInject({ target, propertyKey, identifier });
@@ -20,18 +20,18 @@ export function Autowired(
       attachClassMetadata(
         INJECT_CUSTOM_PARAM,
         {
-          key: 'autowired',
+          key: "autowired",
           parameterIndex,
           propertyKey,
           options: { impl: true },
         },
         target,
         propertyKey,
-        'multi'
+        "multi",
       );
     } else if (isNumber(parameterIndex) && propertyKey) {
       // 保存方法相关信息
-      console.warn('Autowired decorator can not use this method parameter');
+      console.warn("Autowired decorator can not use this method parameter");
     }
   };
 }
