@@ -2,7 +2,7 @@ import {
   APPLICATION_CONTEXT_KEY,
   CONFIG_KEY,
   createCustomPropertyDecorator,
-} from './decorator.manager';
+} from "./decorator.manager";
 
 export function Config(identifier?: string): PropertyDecorator {
   return createCustomPropertyDecorator(CONFIG_KEY, {
