@@ -1,9 +1,9 @@
-export * from './autowired.decorator';
-export * from './configuration.decorator';
-export * from './custom.decorator';
-export * from './controller.decorator';
-export * from './decorator.manager';
-export * from './definitions.decorator';
-export * from './event.decorator';
-export * from './provide.decorator';
-export * from './socket.decorator';
+export * from "./autowired.decorator";
+export * from "./configuration.decorator";
+export * from "./custom.decorator";
+export * from "./controller.decorator";
+export * from "./decorator.manager";
+export * from "./definitions.decorator";
+export * from "./event.decorator";
+export * from "./provide.decorator";
+export * from "./socket.decorator";
