@@ -1,6 +1,6 @@
-import { ObjectIdentifier } from '../interface';
-import { saveModule, SOCKET_KEY } from './decorator.manager';
-import { Singleton } from './definitions.decorator';
+import type { ObjectIdentifier } from "../interface";
+import { saveModule, SOCKET_KEY } from "./decorator.manager";
+import { Singleton } from "./definitions.decorator";
 
 export const Socket = (identifier?: ObjectIdentifier): ClassDecorator => {
   return (target: any) => {
