@@ -1,4 +1,4 @@
-export * from './function.definition';
-export * from './object.creator';
-export * from './object.definition';
-export * from './properties.definition';
+export * from "./function.definition";
+export * from "./object.creator";
+export * from "./object.definition";
+export * from "./properties.definition";
