@@ -1,8 +1,8 @@
-export * from './aspect.service';
-export * from './config.service';
-export * from './decorator.service';
-export * from './environment.service';
-export * from './event.service';
-export * from './ipc.service';
-export * from './lifecycle.service';
-export * from './socket.service';
+export * from "./aspect.service";
+export * from "./config.service";
+export * from "./decorator.service";
+export * from "./environment.service";
+export * from "./event.service";
+export * from "./ipc.service";
+export * from "./lifecycle.service";
+export * from "./socket.service";
