@@ -1,70 +1,70 @@
-import { Init, Singleton } from '../decorators/definitions.decorator';
-import { ILogger, LoggerFactory } from '@electron-boot/logger';
-import { Autowired } from '../decorators/autowired.decorator';
-import {
+import { Init, Singleton } from "../decorators/definitions.decorator";
+import type { ILogger } from "@electron-boot/logger";
+import { LoggerFactory } from "@electron-boot/logger";
+import { Autowired } from "../decorators/autowired.decorator";
+import type {
   IApplicationContext,
   ILifeCycle,
   IObjectLifeCycle,
-} from '../interface';
-import { listModule, CONFIGURATION_KEY } from '../decorators/decorator.manager';
-import { ConfigService } from './config.service';
-import { SocketService } from './socket.service';
-import { FunctionalConfiguration } from '../functional/configuration';
+} from "../interface";
+import { CONFIGURATION_KEY, listModule } from "../decorators/decorator.manager";
+import type { ConfigService } from "./config.service";
+import type { SocketService } from "./socket.service";
+import { FunctionalConfiguration } from "../functional/configuration";
 
 @Singleton()
 export class LifecycleService {
   private logger: ILogger = LoggerFactory.getLogger(LifecycleService);
 
   @Autowired()
-  protected configService: ConfigService;
+  protected configService!: ConfigService;
 
   @Autowired()
-  protected socketService: SocketService;
+  protected socketService!: SocketService;
 
   constructor(readonly applicationContext: IApplicationContext) {}
 
   @Init()
-  protected async init() {
+  protected async init(): Promise<void> {
     // run lifecycle
     const cycles = listModule(CONFIGURATION_KEY) as Array<{
       target: any;
       instance?: any;
     }>;
-
     this.logger.debug(`[core]: Found Configuration length = ${cycles.length}`);
 
-    const lifecycleInstanceList = [];
+    const lifecycleInstanceList: any[] = [];
     for (const cycle of cycles) {
       // 普通类写法
-      this.logger.debug('[core]: Lifecycle init');
+      this.logger.debug(`[core]: Lifecycle init`);
       cycle.instance = await this.applicationContext.getAsync<ILifeCycle>(
         cycle.target,
-        [this.applicationContext]
+        [this.applicationContext],
       );
       cycle.instance && lifecycleInstanceList.push(cycle);
     }
 
     // bind object lifecycle
     await Promise.all([
-      this.runObjectLifeCycle(lifecycleInstanceList, 'onBeforeObjectCreated'),
-      this.runObjectLifeCycle(lifecycleInstanceList, 'onObjectCreated'),
-      this.runObjectLifeCycle(lifecycleInstanceList, 'onObjectInit'),
-      this.runObjectLifeCycle(lifecycleInstanceList, 'onBeforeObjectDestroy'),
+      this.runObjectLifeCycle(lifecycleInstanceList, "onBeforeObjectCreated"),
+      this.runObjectLifeCycle(lifecycleInstanceList, "onObjectCreated"),
+      this.runObjectLifeCycle(lifecycleInstanceList, "onObjectInit"),
+      this.runObjectLifeCycle(lifecycleInstanceList, "onBeforeObjectDestroy"),
     ]);
 
     // exec onConfigLoad()
     await this.runApplicationContextLifeCycle(
       lifecycleInstanceList,
-      'onConfigLoad',
-      configData => {
+      "onConfigLoad",
+      (configData) => {
         if (configData) {
           this.configService.addObject(configData);
         }
-      }
+      },
     );
 
     // exec onReady()
-    await this.runApplicationContextLifeCycle(lifecycleInstanceList, 'onReady');
+    await this.runApplicationContextLifeCycle(lifecycleInstanceList, "onReady");
 
     // exec runtime.run()
     await this.socketService.run();
@@ -72,11 +72,11 @@ export class LifecycleService {
     // exec onSocketReady()
     await this.runApplicationContextLifeCycle(
       lifecycleInstanceList,
-      'onSocketReady'
+      "onSocketReady",
     );
   }
 
-  public async stop() {
+  public async stop(): Promise<void> {
     // stop lifecycle
     const cycles = listModule(CONFIGURATION_KEY) || [];
 
@@ -90,7 +90,7 @@ export class LifecycleService {
         inst = await this.applicationContext.getAsync<ILifeCycle>(cycle.target);
       }
 
-      await this.runApplicationContextLifeCycle(inst, 'onStop');
+      await this.runApplicationContextLifeCycle(inst, "onStop");
     }
 
     // stop socket
@@ -100,16 +100,16 @@ export class LifecycleService {
   private async runApplicationContextLifeCycle(
     lifecycleInstanceOrList: any[] | IObjectLifeCycle,
     lifecycle: string,
-    resultHandler?: (result: any) => void
+    resultHandler?: (result: any) => void,
   ) {
     if (Array.isArray(lifecycleInstanceOrList)) {
       for (const cycle of lifecycleInstanceOrList) {
-        if (typeof cycle.instance[lifecycle] === 'function') {
+        if (typeof cycle.instance[lifecycle] === "function") {
           this.logger.debug(
-            `[core]: Lifecycle run ${cycle.instance.constructor.name} ${lifecycle}`
+            `[core]: Lifecycle run ${cycle.instance.constructor.name} ${lifecycle}`,
           );
           const result = await cycle.instance[lifecycle](
-            this.applicationContext
+            this.applicationContext,
           );
           if (resultHandler) {
             resultHandler(result);
@@ -117,12 +117,12 @@ export class LifecycleService {
         }
       }
     } else {
-      if (typeof lifecycleInstanceOrList[lifecycle] === 'function') {
+      if (typeof lifecycleInstanceOrList[lifecycle] === "function") {
         this.logger.debug(
-          `[core]: Lifecycle run ${lifecycleInstanceOrList.constructor.name} ${lifecycle}`
+          `[core]: Lifecycle run ${lifecycleInstanceOrList.constructor.name} ${lifecycle}`,
         );
         const result = await lifecycleInstanceOrList[lifecycle](
-          this.applicationContext
+          this.applicationContext,
         );
         if (resultHandler) {
           resultHandler(result);
@@ -133,15 +133,15 @@ export class LifecycleService {
 
   private async runObjectLifeCycle(
     lifecycleInstanceList: any[],
-    lifecycle: string
+    lifecycle: string,
   ) {
     for (const cycle of lifecycleInstanceList) {
-      if (typeof cycle.instance[lifecycle] === 'function') {
+      if (typeof cycle.instance[lifecycle] === "function") {
         this.logger.debug(
-          `[core]: Lifecycle run ${cycle.instance.constructor.name} ${lifecycle}`
+          `[core]: Lifecycle run ${cycle.instance.constructor.name} ${lifecycle}`,
         );
         return this.applicationContext[lifecycle](
-          cycle.instance[lifecycle].bind(cycle.instance)
+          cycle.instance[lifecycle].bind(cycle.instance),
         );
       }
     }
