@@ -1,43 +1,44 @@
-import { Init, Singleton } from '../decorators/definitions.decorator';
-import { Autowired } from '../decorators/autowired.decorator';
-import { AspectService } from './aspect.service';
-import { ConfigService } from './config.service';
-import { DecoratorService } from './decorator.service';
-import { IApplicationContext, ISocket } from '../interface';
+import { Init, Singleton } from "../decorators/definitions.decorator";
+import { Autowired } from "../decorators/autowired.decorator";
+import type { AspectService } from "./aspect.service";
+import type { ConfigService } from "./config.service";
+import type { DecoratorService } from "./decorator.service";
+import type { IApplicationContext, ISocket } from "../interface";
 import {
   getProviderUUId,
   listModule,
   SOCKET_KEY,
-} from '../decorators/decorator.manager';
-import { ILogger, LoggerFactory } from '@electron-boot/logger';
+} from "../decorators/decorator.manager";
+import type { ILogger } from "@electron-boot/logger";
+import { LoggerFactory } from "@electron-boot/logger";
 
 @Singleton()
 export class SocketService {
   private logger: ILogger = LoggerFactory.getLogger(SocketService);
   @Autowired()
-  aspectService: AspectService;
+  aspectService!: AspectService;
 
   @Autowired()
-  configService: ConfigService;
+  configService!: ConfigService;
 
   @Autowired()
-  decoratorService: DecoratorService;
+  decoratorService!: DecoratorService;
 
   constructor(readonly applicationContext: IApplicationContext) {}
 
   private globalSocketList: Array<ISocket> = [];
   @Init()
-  async init() {
+  async init(): Promise<void> {
     const sockets: Array<new (...args) => any> = listModule(SOCKET_KEY);
     if (sockets.length) {
       for (const socket of sockets) {
         if (
           !this.applicationContext.registry.hasDefinition(
-            getProviderUUId(socket)
+            getProviderUUId(socket),
           )
         ) {
           this.logger.debug(
-            `[core]: Found socket "${socket.name}" but missing definition, skip initialize.`
+            `[core]: Found socket "${socket.name}" but missing definition, skip initialize.`,
           );
           continue;
         }
@@ -48,24 +49,24 @@ export class SocketService {
 
     await this.aspectService.loadAspect();
   }
-  async run() {
+  async run(): Promise<void> {
     for (const socket of this.globalSocketList) {
       // if enabled, just run socket
       if (socket.isEnable()) {
         // app init
         await socket.run();
         this.logger.debug(
-          `[core]: Found socket "${socket.getName()}" and run.`
+          `[core]: Found socket "${socket.getName()}" and run.`,
         );
       }
     }
   }
 
-  async stop() {
+  async stop(): Promise<void> {
     await Promise.all(
-      Array.from(this.globalSocketList).map(socketInstance => {
+      Array.from(this.globalSocketList).map((socketInstance) => {
         return socketInstance.stop();
-      })
+      }),
     );
   }
 }
