@@ -1,2 +1,2 @@
-export const ORIGIN_ARGS = Symbol('origin_args')
-export const ORIGIN_ERROR = Symbol('origin_error')
+export const ORIGIN_ARGS = Symbol("origin_args");
+export const ORIGIN_ERROR = Symbol("origin_error");
